# dashboard/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Flaglane API",
    "description" : "Feature flags and remote configuration. `/api/**` serves the dashboard with a JWT; `/sdk/**` serves SDKs with an API key.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "0.1.0"
  },
...
  "paths" : {
...
    "/api/projects/{projectKey}/environments/{envKey}/keys" : {
      "get" : {
        "tags" : [ "API keys" ],
        "summary" : "List an environment's keys",
        "description" : "Prefix and metadata only.",
        "operationId" : "list_3",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The environment's keys, revoked ones included",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ApiKeyResponse"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or key among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "API keys" ],
        "summary" : "Issue a key",
        "description" : "The response carries the key, and is the only response that ever will (FR-KEY-002). Store it now; Flaglane keeps only a hash.",
        "operationId" : "issue",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IssueApiKeyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Issued",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuedApiKeyResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or key among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/projects/{projectKey}/flags/{flagKey}/config/{envKey}" : {
      "get" : {
        "tags" : [ "Flags" ],
        "summary" : "Get a flag's configuration in an environment",
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The configuration",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagConfigResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Flags" ],
        "summary" : "Change a flag's configuration in an environment",
        "description" : "Kill switch, fallthrough value, rollout percentage and rollout salt. Takes effect on the environment's next ruleset. **Changing `rolloutSalt` re-buckets every user of this flag.**",
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFlagConfigRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "The flag is archived",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "The configuration as changed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagConfigResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/projects/{projectKey}/environments/{envKey}/keys/{keyId}" : {
      "delete" : {
        "tags" : [ "API keys" ],
        "summary" : "Revoke a key",
        "description" : "Refused on every SDK request from the moment this returns (FR-KEY-003). Revoking a revoked key is accepted and changes nothing.",
        "operationId" : "revoke",
        "parameters" : [ {
          "name" : "keyId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Revoked"
          },
          "404" : {
            "description" : "No such project, environment or key among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "format" : "uri"
          },
          "properties" : {
            "type" : [ "object", "null" ],
            "additionalProperties" : { }
          }
        }
      },
...
      "IssueApiKeyRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "checkout-service",
            "maxLength" : 100,
            "minLength" : 0
          },
          "type" : {
            "type" : "string",
            "description" : "`server` reads every flag; `client` reads client-side-visible flags only, with no user overrides",
            "enum" : [ "server", "client" ]
          }
        },
        "required" : [ "name", "type" ]
      },
      "IssuedApiKeyResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "server", "client" ]
          },
          "prefix" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "key" : {
            "type" : "string",
            "description" : "The key. Shown once and never again",
            "example" : "flg_srv_Xk3mP9qaR2vT8wY1zB4cD6eF0gH5jK7lM9nQ2sU4xW6"
          }
        },
        "required" : [ "createdAt", "id", "key", "name", "prefix", "type" ]
      },
...
      "UpdateFlagConfigRequest" : {
        "type" : "object",
        "properties" : {
          "enabled" : {
            "type" : [ "boolean", "null" ],
            "description" : "The kill switch. Disabled returns false to everyone"
          },
          "fallthroughValue" : {
            "type" : [ "boolean", "null" ],
            "description" : "Returned when the flag is enabled and nothing else matched. While true, the rollout has no effect: everyone it does not reach gets true anyway"
          },
          "rolloutPercentage" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Whole percent of users, stored as basis points (ADR-011)",
            "maximum" : 100,
            "minimum" : 0
          },
          "rolloutSalt" : {
            "type" : [ "string", "null" ],
            "description" : "Changing the salt re-buckets every user of this flag: who has it at a given percentage changes. Give flags the same salt to roll them out to the same users (ADR-010)",
            "pattern" : "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
          }
        }
      },
      "FlagConfigResponse" : {
        "type" : "object",
        "properties" : {
          "flagKey" : {
            "type" : "string",
            "example" : "new-checkout"
          },
          "environment" : {
            "type" : "string",
            "example" : "production"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "offValue" : {
            "type" : "boolean",
            "description" : "Always false in v0.x (ADR-009)"
          },
          "fallthroughValue" : {
            "type" : "boolean"
          },
          "rolloutPercentage" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0
          },
          "rolloutSalt" : {
            "type" : "string",
            "example" : "new-checkout"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "enabled", "environment", "fallthroughValue", "flagKey", "offValue", "rolloutPercentage", "rolloutSalt", "updatedAt" ]
      },
      "ApiKeyResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string",
            "example" : "checkout-service"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "server", "client" ]
          },
          "prefix" : {
            "type" : "string",
            "description" : "Enough to tell keys apart; not secret",
            "example" : "flg_srv_Xk3mP9qa"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastUsedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "Accurate to about a minute"
          },
          "revokedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "id", "lastUsedAt", "name", "prefix", "revokedAt", "type" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "An access token from `POST /api/auth/login`, valid for 8 hours.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "sdk-key" : {
        "type" : "http",
        "description" : "An API key issued for one environment. A server key reads every flag; a client key reads client-side-visible flags only, with no user overrides.",
        "scheme" : "bearer",
        "bearerFormat" : "flg_srv_… or flg_cli_…"
      }
    }
  }
}